Restructure
Restructure the project file structure.
Created a Mp2tClipImpl sub-project that will build a library where most of the core code will be located.
