+++ Mp2tClipImpl/include/Mp2tClipImpl/Clipper.h
